expose base geopose prop in navigation map
