+++ packages/blocks/src/lib/index.ts
@@ -1,9 +1,21 @@
+// MapLibre components
 export { default as MapLibre } from './maplibre/index.svelte';
 export { default as MapLibreMarker } from './maplibre/marker.svelte';
 export { useMapLibre, useMapLibreEvent } from './maplibre/hooks';
+export type { LngLat, GeoPose, Waypoint } from './maplibre/types';
 
+// Slam components
 export { default as SlamMap2D } from './slam-map-2d/index.svelte';
 
+// Navigation map components
 export { default as NavigationMap } from './navigation-map/index.svelte';
-export type * from './navigation-map/types';
-export { NavigationTab } from './navigation-map/types';
+export type {
+  NavigationTab,
+  NavigationTabType,
+  Shapes,
+  CapsuleGeometry,
+  SphereGeometry,
+  BoxGeometry,
+  Geometry,
+  Obstacle,
+} from './navigation-map/types';
